Python_I/Chrome_Password_Decryptor/requirements.txt
pycryptodomex
pypiwin32